perf: extract factory-core.ts for lightweight namespace help
Extract defineGroup, configureJsonHelp, hideBlockedCommands,
isCommandAllowed, stripTransportMeta, and all shared type exports into
factory-core.ts (66 source lines, ~91 compiled lines).

cli.ts, namespaces.ts, and cloud/register-lazy.ts import from
factory-core.ts instead of factory.ts. This means cloud --help and bare
startup never load factory.ts (453 compiled lines with defineCommand,
buildLeafHandle, output rendering, etc.).

factory.ts imports and re-exports from factory-core.ts so existing
consumers (es/register.ts, tests) continue to work unchanged and test
coverage flows through.

Metric impact: cloud --help skips 362 lines of JS parsing (~3-5ms).
Running total: 606ms → ~137ms
