+++ test/cloud/register-lazy.test.ts
@@ -3,25 +3,12 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { describe, it, before, after } from 'node:test'
+import { describe, it } from 'node:test'
 import assert from 'node:assert/strict'
 import { registerCloudCommandsLazy } from '../../src/cloud/register-lazy.ts'
 
 describe('registerCloudCommandsLazy', () => {
-  const ORIGINAL_ARGV = process.argv.slice()
-
-  before(() => {
-    // Simulate `elastic cloud --help` (top-level, no sub-group)
-    process.argv.splice(2)
-    process.argv.push('cloud', '--help')
-  })
-
-  after(() => {
-    process.argv.splice(0)
-    ORIGINAL_ARGV.forEach(a => process.argv.push(a))
-  })
-
-  it('returns a top-level "cloud" group', async () => {
+  it('returns a top-level "cloud" group (no sub-namespace)', async () => {
     const group = await registerCloudCommandsLazy()
     assert.equal(group.name(), 'cloud')
   })
@@ -34,13 +21,9 @@ describe('registerCloudCommandsLazy', () => {
     assert.ok(names.includes('users'), 'should have users')
   })
 
-  it('builds real tree when a sub-group is targeted', async () => {
-    // Simulate `elastic cloud hosted --help`
-    process.argv.splice(2)
-    process.argv.push('cloud', 'hosted', '--help')
-    const group = await registerCloudCommandsLazy()
+  it('builds real tree when targetSubNamespace is set', async () => {
+    const group = await registerCloudCommandsLazy('hosted')
     assert.equal(group.name(), 'cloud')
-    // Real tree has actual sub-commands under hosted
     const hosted = group.commands.find(c => c.name() === 'hosted')
     assert.ok(hosted != null, 'should have hosted group')
     assert.ok(hosted.commands.length > 0, 'hosted should have sub-commands in real tree')
